requirements.txt
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pyTelegramBotAPI==0.3.0
pytest==7.1.2
requests==2.28.1
six==1.16.0
tomli==2.0.1
torch==1.12.0
torchvision==0.13.0
typing-extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.2.0
zipp==3.8.1
